+++ cli/internal/hasura/commonmetadata/commonmetadata_test.go
@@ -59,6 +59,7 @@ func TestClient_ExportMetadata(t *testing.T) {
       "tables": [],
       "configuration": {
         "connection_info": {
+          "use_prepared_statements": true,
           "database_url": {
             "from_env": "HASURA_GRAPHQL_DATABASE_URL"
           },
@@ -66,8 +67,7 @@ func TestClient_ExportMetadata(t *testing.T) {
             "retries": 1,
             "idle_timeout": 180,
             "max_connections": 50
-          },
-	  "use_prepared_statements": true
+          }
         }
       }
     }