+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Add entries here in the order of: server, console, cli, docs, others)
 
 - console: add bigquery support (#1000)
+- cli: add support for bigquery in metadata operations
 
 ## v2.0.0-alpha.8
 
@@ -54,6 +55,7 @@ In the future, we will probably offer a way to explicitly choose which behaviour
 - console: add details button to the success notification to see inserted row
 - console: add request preview for REST endpoints
 - cli: fix errors being ignored during `metadata apply` in config v3 (fix #6784)
+
 
 ## v2.0.0-alpha.7